Create User APi done

+++ MensFashion/src/main/java/com/mensfashion/dto/UserDTO.java
@@ -14,7 +14,7 @@ public class UserDTO {
 	private String name;
 	private String email;
 	private String password;
-	private BigDecimal phoneNum;
+	private long phoneNum;
 	
 	@Override
 	public String toString() {
